+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/asserts/segment/limit/LimitClauseAssert.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.integrate.asserts.segment.pagination;
+package org.apache.shardingsphere.sql.parser.integrate.asserts.segment.limit;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
